[FIX] split keychain backend model and test into 2 files
