+++ keychain/models/__init__.py
@@ -1,1 +1,2 @@
 from . import keychain
+from . import keychain_backend
