+++ keychain/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_keychain
+from . import test_keychain_backend
